# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse:
    properties:
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image_url:
        example: https://example.com/cat.jpg
        type: string
      name:
        example: Kaos Polos
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.FabricColorRequest:
    properties:
      hex_code:
        example: '#4b5320'
        type: string
      name:
        example: Olive
        type: string
    required:
    - hex_code
    - name
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.FabricColorResponse:
    properties:
      hex_code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductCreateRequest:
    properties:
      base_price:
        example: 185000
        type: number
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      description:
        example: Kemeja taktikal bahan ripstop cotton 65/35
        type: string
      design_model:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelRequest'
      fabric_summary:
        example: Ripstop
        type: string
      fabrics:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricRequest'
        type: array
      gsm_info:
        example: 210gsm
        type: string
      image_urls:
        example:
        - '["https://supa.../1.jpg"]'
        items:
          type: string
        type: array
      key_features:
        example:
        - '["Bahan ripstop anti robek"'
        - ' "Dual chest pocket velcro"]'
        items:
          type: string
        type: array
      name:
        example: Kemeja Taktikal Premium 7200
        type: string
      slug:
        example: kemeja-taktikal-premium-7200
        type: string
      wholesale:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceRequest'
        type: array
    required:
    - base_price
    - category_id
    - name
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricRequest:
    properties:
      base_price:
        example: 185000
        minimum: 0
        type: number
      care_instruction:
        example: Cuci mesin air dingin, jangan diputihkan
        type: string
      colors:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.FabricColorRequest'
        type: array
      composition:
        example: 65% Cotton / 35% Polyester
        type: string
      description:
        example: Kuat & anti robek, 210gsm
        type: string
      is_default:
        example: true
        type: boolean
      name:
        example: Ripstop Cotton 65/35
        type: string
      price_adjustment:
        example: 0
        type: number
    required:
    - name
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricResponse:
    properties:
      base_price:
        type: number
      care_instruction:
        type: string
      colors:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.FabricColorResponse'
        type: array
      composition:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      price_adjustment:
        type: number
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductImageResponse:
    properties:
      id:
        type: string
      image_url:
        type: string
      is_primary:
        type: boolean
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelRequest:
    properties:
      description:
        example: Template kemeja taktikal 2 saku
        type: string
      name:
        example: Series 1 — Lengan Panjang
        type: string
      type:
        example: long_sleeve
        type: string
      views:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelViewRequest'
        type: array
    required:
    - name
    - type
    - views
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      views:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelViewResponse'
        type: array
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelViewRequest:
    properties:
      art_url:
        example: https://supa.../front-art.png
        type: string
      height:
        example: 1920
        type: integer
      mask_url:
        example: https://supa.../front-mask.png
        type: string
      side:
        enum:
        - front
        - back
        example: front
        type: string
      width:
        example: 1756
        type: integer
    required:
    - art_url
    - mask_url
    - side
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelViewResponse:
    properties:
      art_url:
        type: string
      height:
        type: integer
      id:
        type: string
      mask_url:
        type: string
      side:
        type: string
      width:
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductResponse:
    properties:
      base_price:
        example: 185000
        type: number
      category:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse'
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      created_at:
        type: string
      description:
        example: Kemeja taktikal bahan ripstop cotton
        type: string
      design_model:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelResponse'
      fabric_summary:
        example: Ripstop
        type: string
      fabrics:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricResponse'
        type: array
      gsm_info:
        example: 210gsm
        type: string
      id:
        example: 999e4567-e89b-12d3-a456-426614174000
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductImageResponse'
        type: array
      key_features:
        items:
          type: string
        type: array
      name:
        example: Kemeja Taktikal Premium 7200
        type: string
      rating:
        example: 4.9
        type: number
      review_count:
        example: 318
        type: integer
      slug:
        example: kemeja-taktikal-premium-7200
        type: string
      sold_count:
        example: 1240
        type: integer
      updated_at:
        type: string
      wholesale:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceResponse'
        type: array
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductUpdateRequest:
    properties:
      base_price:
        example: 190000
        type: number
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      description:
        example: Versi update
        type: string
      design_model:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelRequest'
      fabric_summary:
        example: Ripstop Cotton
        type: string
      fabrics:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricRequest'
        type: array
      gsm_info:
        example: 220gsm
        type: string
      image_urls:
        items:
          type: string
        type: array
      key_features:
        items:
          type: string
        type: array
      name:
        example: Kemeja Taktikal Premium 7200 V2
        type: string
      slug:
        example: kemeja-taktikal-premium-7200-v2
        type: string
      wholesale:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceRequest'
        type: array
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceRequest:
    properties:
      fabric_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      max_qty:
        example: 24
        type: integer
      min_qty:
        example: 6
        type: integer
      unit_price:
        example: 175000
        type: number
    required:
    - min_qty
    - unit_price
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceResponse:
    properties:
      fabric_id:
        type: string
      id:
        type: string
      max_qty:
        type: integer
      min_qty:
        type: integer
      unit_price:
        type: number
    type: object
  github_com_faridlan_sikon-api_internal_domain.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_utils.EmptyObj:
    type: object
  github_com_faridlan_sikon-api_internal_utils.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        example: pesan error
        type: string
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_domain.PaginationMeta'
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductResponse'
      message:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj:
    properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.EmptyObj'
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk MVP ERP SIKOn.
  title: SIKOn API (Sistem Integrasi Konveksi Online)
  version: "1.0"
paths:
...
  /products:
    get:
      description: Mengambil daftar seluruh produk dengan pagination, filter kategori,
        harga, dan sorting
      parameters:
      - default: 1
        description: Nomor Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Batas Data per Halaman
        in: query
        name: limit
        type: integer
      - description: Pencarian nama/deskripsi
        in: query
        name: search
        type: string
      - description: Filter Kategori ID (UUID)
        in: query
        name: category_id
        type: string
      - description: Filter Harga Minimal
        format: float64
        in: query
        name: min_price
        type: number
      - description: Filter Harga Maksimal
        format: float64
        in: query
        name: max_price
        type: number
      - description: Urutan Data (popular, price_low, price_high, newest)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Menambahkan produk baru ke dalam katalog beserta opsi kain, warna,
        grosir, dan template canvas
      parameters:
      - description: Data Produk Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Menghapus data produk secara permanen
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Product
      tags:
      - Products
    get:
      description: Mengambil detail data produk berdasarkan ID
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Product By ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Memperbarui data produk (kain, warna, grosir, dan template canvas)
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Update Produk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Product
      tags:
      - Products
  /products/slug/{slug}:
    get:
      description: Mengambil detail data produk berdasarkan Slug URL
      parameters:
      - description: Product Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      summary: Get Product By Slug
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"